refactor: infrastracture 패키지 순환참조 제거
