+++ src/main/java/kitchenpos/dao/infrastructure/JdbcTemplateMenuDao.java
@@ -1,12 +1,11 @@
-package kitchenpos.infrastructure;
+package kitchenpos.dao.infrastructure;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.Menu;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -16,14 +15,14 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class JdbcTemplateTableGroupDao {
-    private static final String TABLE_NAME = "table_group";
+public class JdbcTemplateMenuDao {
+    private static final String TABLE_NAME = "menu";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateTableGroupDao(DataSource dataSource) {
+    public JdbcTemplateMenuDao(DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -31,13 +30,13 @@ public class JdbcTemplateTableGroupDao {
         ;
     }
 
-    public TableGroup save(TableGroup entity) {
+    public Menu save(Menu entity) {
         SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
-    public Optional<TableGroup> findById(Long id) {
+    public Optional<Menu> findById(Long id) {
         try {
             return Optional.of(select(id));
         } catch (EmptyResultDataAccessException e) {
@@ -45,22 +44,31 @@ public class JdbcTemplateTableGroupDao {
         }
     }
 
-    public List<TableGroup> findAll() {
-        String sql = "SELECT id, created_date FROM table_group";
+    public List<Menu> findAll() {
+        String sql = "SELECT id, name, price, menu_group_id FROM menu ";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private TableGroup select(Long id) {
-        String sql = "SELECT id, created_date FROM table_group WHERE id = (:id)";
+    public long countByIdIn(List<Long> ids) {
+        String sql = "SELECT COUNT(*) FROM menu WHERE id IN (:ids)";
+        SqlParameterSource parameters = new MapSqlParameterSource()
+                .addValue("ids", ids);
+        return jdbcTemplate.queryForObject(sql, parameters, Long.class);
+    }
+
+    private Menu select(Long id) {
+        String sql = "SELECT id, name, price, menu_group_id FROM menu WHERE id = (:id)";
         SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private TableGroup toEntity(ResultSet resultSet) throws SQLException {
-        return new TableGroup(
-                resultSet.getLong(KEY_COLUMN_NAME),
-                resultSet.getObject("created_date", LocalDateTime.class)
+    private Menu toEntity(ResultSet resultSet) throws SQLException {
+        return new Menu(
+                resultSet.getLong("id"),
+                resultSet.getString("name"),
+                resultSet.getBigDecimal("price"),
+                resultSet.getLong("menu_group_id")
         );
     }
 }
