+++ src/main/java/kitchenpos/dao/infrastructure/JdbcTemplateTableGroupDao.java
@@ -1,12 +1,12 @@
-package kitchenpos.infrastructure;
+package kitchenpos.dao.infrastructure;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.TableGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -16,14 +16,14 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class JdbcTemplateMenuGroupDao implements MenuGroupDao {
-    private static final String TABLE_NAME = "menu_group";
+public class JdbcTemplateTableGroupDao {
+    private static final String TABLE_NAME = "table_group";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateMenuGroupDao(DataSource dataSource) {
+    public JdbcTemplateTableGroupDao(DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -31,15 +31,13 @@ public class JdbcTemplateMenuGroupDao implements MenuGroupDao {
         ;
     }
 
-    @Override
-    public MenuGroup save(MenuGroup entity) {
+    public TableGroup save(TableGroup entity) {
         SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
-    @Override
-    public Optional<MenuGroup> findById(Long id) {
+    public Optional<TableGroup> findById(Long id) {
         try {
             return Optional.of(select(id));
         } catch (EmptyResultDataAccessException e) {
@@ -47,28 +45,22 @@ public class JdbcTemplateMenuGroupDao implements MenuGroupDao {
         }
     }
 
-    @Override
-    public List<MenuGroup> findAll() {
-        String sql = "SELECT id, name FROM menu_group";
+    public List<TableGroup> findAll() {
+        String sql = "SELECT id, created_date FROM table_group";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    @Override
-    public boolean existsById(Long id) {
-        String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END FROM menu_group WHERE id = (:id)";
-        SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("id", id);
-        return jdbcTemplate.queryForObject(sql, parameters, Boolean.class);
-    }
-
-    private MenuGroup select(Long id) {
-        String sql = "SELECT id, name FROM menu_group WHERE id = (:id)";
+    private TableGroup select(Long id) {
+        String sql = "SELECT id, created_date FROM table_group WHERE id = (:id)";
         SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private MenuGroup toEntity(ResultSet resultSet) throws SQLException {
-        return new MenuGroup(resultSet.getLong("id"), resultSet.getString("name"));
+    private TableGroup toEntity(ResultSet resultSet) throws SQLException {
+        return new TableGroup(
+                resultSet.getLong(KEY_COLUMN_NAME),
+                resultSet.getObject("created_date", LocalDateTime.class)
+        );
     }
 }
